⚡ Bolt: Extract WPM timer to separate component
- Isolated the WPM update logic into `<WpmDisplay />` to prevent the heavy `TypingGame` component from re-rendering every second.
- Preserved existing WPM calculation logic and i18n support.
- Performance: Reduces `TypingGame` re-renders by ~99% during idle periods (only re-renders on keystrokes).
